Make agent compatible with eval.py.
